requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
langchain>=0.1.0
langchain-openai>=0.0.8
Flask-Limiter==3.11.0
Flask-Talisman==1.1.0
gunicorn==23.0.0
limits==4.2
ordered-set==4.1.0
openai>=1.12.0
Werkzeug==3.0.1
langgraph>=0.0.26
langgraph-checkpoint>=2.0.10 